requirements.txt
beautifulsoup4
requests
urllib3
robotexclusionrulesparser
tqdm
pandas
scikit-learn
joblib
nltk
streamlit
python-dotenv
tensorflow 
pandas 
scikit-learn 
joblib 
streamlit 
langchain-nvidia-ai-endpoints 
langchain 
python-dotenv 
tqdm
# duckduckgo-search
ddgs
langchain_community
langchain_tavily